% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance_tables.R
\name{abundance_table}
\alias{abundance_table}
\alias{abundance_table,transcript_quantifier-method}
\title{View abundances}
\usage{
abundance_table(tq)

\S4method{abundance_table}{transcript_quantifier}(tq)
}
\arguments{
\item{tq}{A \link{transcript_quantifier-class} object}
}
\description{
Produces a table with estimated trancript abundances for a given
\code{\link{transcript_quantifier-class}} object
}
