% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{add_data}
\alias{add_data}
\alias{add_data,transcript_quantifier-method}
\title{Add data}
\usage{
add_data(tq, bigwig_plus = NULL, bigwig_minus = NULL)

\S4method{add_data}{transcript_quantifier}(tq, bigwig_plus = NULL, bigwig_minus = NULL)
}
\arguments{
\item{tq}{A \link{transcript_quantifier-class} object}

\item{bigwig_plus}{the path to a bigwig for reads on the plus strand}

\item{bigwig_minus}{the path to a bigwig for reads on the minus strand}
}
\value{
A \link{transcript_quantifier-class} object with count data added.
}
\description{
adds data from bigwig to a \link{transcript_quantifier-class}
object
}
