% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{apply_bigwig_seqinfo}
\alias{apply_bigwig_seqinfo}
\title{Apply seqinfo from bigwig to GRanges}
\usage{
apply_bigwig_seqinfo(
  x,
  bigwig_file,
  drop_unused = TRUE,
  pruning_mode = c("error", "coarse", "fine", "tidy")
)
}
\arguments{
\item{x}{an object trhat has a \code{seqinfo} method (e.g. GRanges)}

\item{bigwig_file}{a string pointing to a bigwig file}

\item{drop_unused}{boolean drop unused seqlevels}

\item{pruning_mode}{see \code{pruning.mode} from \link[GenomeInfoDb]{seqinfo}}
}
\value{
a named vector of the polymerase ratios
}
\description{
Applies the seqinfo from a bigwig to any other object that has a valid \code{seqinfo}
method matches seqlevels, and prunes out of range entries
}
