% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcript_shape_profile-class.R
\name{apply_shape_profile}
\alias{apply_shape_profile}
\alias{apply_shape_profile,transcript_quantifier,transcript_shape_profile-method}
\title{Change shape of transcript models

Create a \code{\link{transcript_quantifier-class}} that uses the shape
profile from a \code{\link{transcript_shape_profile-class}}}
\usage{
apply_shape_profile(tq, tsp)

\S4method{apply_shape_profile}{transcript_quantifier,transcript_shape_profile}(tq, tsp)
}
\arguments{
\item{tq}{A \code{\link{transcript_quantifier-class}} object}

\item{tsp}{A \link{transcript_shape_profile-class} object}
}
\description{
Change shape of transcript models

Create a \code{\link{transcript_quantifier-class}} that uses the shape
profile from a \code{\link{transcript_shape_profile-class}}
}
