% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bins.R
\name{bin_seq}
\alias{bin_seq}
\title{Create bins for a group of overlapped transcripts}
\usage{
bin_seq(seqname, start, end, strand, bin_size)
}
\arguments{
\item{seqname}{chromsome name}

\item{start}{start (single value)}

\item{end}{end (single value)}

\item{strand}{strand (single value)}

\item{bin_size}{An integer, used to tile the gene region. Default is 250bp.}
}
\value{
A \code{\link[data.table]{data.table}} object, containing
the binned region.
}
\description{
Function for generating bins for a group of overlapped transcripts.
}
