% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txdb_handling.R
\name{cache_defaults}
\alias{cache_defaults}
\title{Create cache defaults}
\usage{
cache_defaults()
}
\value{
A list with two elements \code{dir} and \code{path_file}
}
\description{
Creates paths for default cache_dir and cache_path.txt
file
}
