% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txdb_handling.R
\name{cache_get_dir}
\alias{cache_get_dir}
\title{Get local cache path}
\usage{
cache_get_dir()
}
\description{
Downloaded databases are cached locally. This function
gets the local cache path from the  the R environmental variable
\code{CACHE_DIR_TUSEL} or returns an error if it does not exist.
}
\seealso{
\code{\link{cache_set_dir}}
}
