% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_data.R
\name{cache_set_dir}
\alias{cache_set_dir}
\title{Set local cache path}
\usage{
cache_set_dir(cache_dir = NULL)
}
\arguments{
\item{cache_dir}{Character path to new local files path. If null,
path will be reset to default user data directory location.}
}
\description{
Downloaded databases are cached locally. This function
sets the R environmental variable \code{CACHE_DIR_TUSEL}
with the location of the cache dir. If \code{cache_dir} is
specified, it sets the cache directory to that location,
creating the specified directory if one does not exit.
Otherwise it just uses the previously specified cache dir.
If no cache directory has been created  and \code{cache_dir}
is unspecified it creates one in the default location.
}
\seealso{
\code{\link{cache_get_dir}}
}
