% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_data.R
\name{cache_txdb}
\alias{cache_txdb}
\title{Cache a TxDb}
\usage{
cache_txdb(
  txdb,
  species_name = NULL,
  genome_name = NULL,
  annotation_version = NULL
)
}
\arguments{
\item{txdb}{a TxDb object}

\item{species_name}{the name of the species (defaults to using
the organism name from the txdb if none specified).}

\item{genome_name}{the name of the genome (defaults to using
the BioMart dataset version from the txdb if none specified).}

\item{annotation_version}{a version id (defaults to using the BioMart
database version if none specified)}
}
\description{
Save a txdb in the local cache.
}
\seealso{
\code{\link{cache_get_dir}}
}
