% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tss_predictor.R
\name{collect_tss_features}
\alias{collect_tss_features}
\title{Generate GRangesList of windows}
\usage{
collect_tss_features(
  transcripts,
  bigwig_plus,
  bigwig_minus,
  bins = 21L,
  bin_width = 51L,
  remove_incomplete = TRUE,
  remove_all_zero = FALSE
)
}
\arguments{
\item{transcripts}{a \code{\link[GenomicRanges]{GRanges-class}} object}

\item{bigwig_plus}{the path to a bigwig for reads on the plus strand}

\item{bigwig_minus}{the path to a bigwig for reads on the minus strand}

\item{bins}{a single value or vector of length equal to bin_width indicating
the number of bins to generate, with the central bin centered on the
focal site, must be odd}

\item{bin_width}{a vector of one or more odd valued integers indicating the
width of the bins to be generated. If multiple value are specified, windows are
generated for all sizes specified}

\item{remove_incomplete}{remove regions where any window features go outside
the chromosomal range}

\item{remove_all_zero}{remove TSS that have only 0 values in their feature
vector}
}
\value{
A list of vectors with each one corresponding to one set of bins and
each element of a vector corresponding to a bin
}
\description{
Generates a set of windows for each transcript in a
\code{\link[GenomicRanges]{GRanges-class}} object. Uses the TSS of each
region as the 0th co-ordinate in generating windows.
}
