% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcript_models_build_helpers.R
\name{combine_masks}
\alias{combine_masks}
\title{Function for combining masks}
\usage{
combine_masks(model_masks, add_masks)
}
\arguments{
\item{model_masks, add_masks}{A list of vectors, where each vector indicates
the bins that should be masked.}
}
\value{
A list of vectors, where each vector has combined the masks, and indicates
the bins that should be masked.
}
\description{
Function for combining masks
}
