% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcript_shape_profile-class.R
\name{covered_bases}
\alias{covered_bases}
\title{Gets per-query coverage}
\usage{
covered_bases(query, subject)
}
\arguments{
\item{query}{a \link[GenomicRanges]{GRanges-class}}

\item{subject}{a \link[GenomicRanges]{GRanges-class}}
}
\description{
Returns vector with number of bases overlapped by the subject intervals per
query interval.
}
