% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcript_models_build_helpers.R
\name{create_additional_masks}
\alias{create_additional_masks}
\title{Function for generating masks for additional genomic regions}
\usage{
create_additional_masks(bins, add_mask)
}
\arguments{
\item{bins}{A \code{\link[GenomicRanges]{GRangesList-class}} object,
each element in the list contains the bins for each corresponding group of
transcripts}

\item{add_mask}{A \code{\link[GenomicRanges]{GRanges-class}}
object records genomic regions being masked. If seqnames in add_mask are not
available in transcripts, entries associated with these seqnames will be removed.}
}
\value{
A list of vectors, where each vector indicates the bins that should
be masked
}
\description{
Function for generating masks for additional genomic regions
}
