% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bin_masks.R
\name{create_bin_masks}
\alias{create_bin_masks}
\title{Create data masks}
\usage{
create_bin_masks(txdb, bins, mask_start_bins = 0, mask_end_bins = 0)
}
\arguments{
\item{txdb}{a \code{\link[GenomicFeatures]{Txdb-class}} object}

\item{bins}{a \code{\link[GenomicRanges]{GRangesList-class}} object}

\item{mask_start_bins}{The number of bins at the start of each transcript to
mask}

\item{mask_end_bins}{The number of bins at the end of each transcript to
mask}
}
\value{
A list of bit vectors
}
\description{
Creates a set of bin masks for transcript starts and ends
}
