% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_model_masks.R
\name{create_model_masks}
\alias{create_model_masks}
\title{Create model masks}
\usage{
create_model_masks(
  transcript_models,
  strand,
  mask_start_bins = 0,
  mask_end_bins = 0
)
}
\arguments{
\item{transcript_models}{A list of matricies containing transcript models}

\item{strand}{a vector containing one entry per group of transcript models
indicating the strand}

\item{mask_start_bins}{the number of bins with non-zero entries at the start
of a transcript that should be masked (Used for masking both start and end if
strand is \'*\')}

\item{mask_end_bins}{the number of bins with non-zero entries at the end
of a transcript that should be masked (Ignored if strand is \'*\')}
}
\value{
A list of vectors, where each vector indicates the bins that should
be masked
}
\description{
Function for generating masks
}
