% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_transcript_models.R
\name{create_transcript_models}
\alias{create_transcript_models}
\title{Create transcript models}
\usage{
create_transcript_models(transcript_groups, bins)
}
\arguments{
\item{transcript_groups}{A \code{\link[GenomicRanges]{GRangesList-class}}
object, each element in the list contains transcripts from one gene or
several genes which are overlapped.}

\item{bins}{A \code{\link[GenomicRanges]{GRangesList-class}} object,
each element in the list contains the bins for each corresponding group of
transcripts}
}
\value{
A list of matrices where each row in the matrix corresponds to a
transcript and each column is a bin
}
\description{
Function for generating binned transcript models given a set of bins
}
