% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ensembl_model.R
\name{download_txdb}
\alias{download_txdb}
\title{Download transcript model from biomart}
\usage{
download_txdb(species = NULL, ensembl_version = NULL, cache = TRUE)
}
\arguments{
\item{species}{the species to download the TxDb for. Use the first letter
of the genus and the full species as one string (e.g. homo sapiens -> 
hsapiens)}

\item{ensembl_version}{The Ensembl release version.}
}
\value{
A mart object.
}
\description{
This function creates a TxDb object from the biomaRt ensembl database,
converts it to a TxDb, and returns it.
}
