% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitting.R
\name{fit}
\alias{fit}
\alias{fit,transcript_quantifier-method}
\title{Fit model}
\usage{
fit(
  tq,
  lambda = 0,
  transform = "log",
  inactive_transcripts = NA,
  verbose = FALSE
)

\S4method{fit}{transcript_quantifier}(
  tq,
  lambda = 0,
  transform = "log",
  inactive_transcripts = NA,
  verbose = FALSE
)
}
\arguments{
\item{tq}{A \link{transcript_quantifier-class} object}

\item{lambda}{the weight of the lasso penalty (not implemented at this time)}

\item{transform}{how to transform the data (either "identity" or "log")}

\item{inactive_transcripts}{a character vector listing transcripts for which abundance
values should be fixed to 0. IMPORTANT: In the case where multiple transcripts are
assigned to a single model (due to identical models at the specified bin scale) this
will be overridden if one or more transcripts assigned to the same model are active.}

\item{verbose}{if TRUE shows progress bar for fitting (default: FALSE)}
}
\description{
Estimates trancript abundances for a given
\code{\link{transcript_quantifier-class}} object under a lasso penalty
}
