% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tq.R
\name{get_abundance}
\alias{get_abundance}
\title{retrieve abundance for plotting}
\usage{
get_abundance(transcript_quantifier, chrom, start, end)
}
\arguments{
\item{transcript_quantifier}{A \link{transcript_quantifier-class} object}

\item{chrom}{the chromosome of your query region (single value)}

\item{start}{the start position of your query region (single value)}

\item{end}{the end position of your query region (single value)}
}
\value{
a \link[GenomicRanges]{GRanges-class} objects with the adbundance in
the metadate column(s).
}
\description{
retrieve abundance for plotting
}
