% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tq.R
\name{get_data}
\alias{get_data}
\alias{get_data,transcript_quantifier-method}
\alias{get_data,character-method}
\title{Get data}
\usage{
get_data(data_source, chrom, start, end)

\S4method{get_data}{transcript_quantifier}(data_source, chrom, start, end)

\S4method{get_data}{character}(data_source, chrom, start, end)
}
\arguments{
\item{data_source}{a \link{transcript_quantifier-class} object or string
giving the path to a bigwig file}

\item{chrom}{the chromosome of your query region (single value)}

\item{start}{the start position of your query region (single value)}

\item{end}{the end position of your query region (single value)}
}
\value{
a \link[GenomicRanges]{GRanges-class} objects with the read counts in
the \code{value} column
}
\description{
Retrives data for plotting. Returned ranges may go a bit outside
the queried ones if the overlapping bins do not perfectly align with the
user provided boundries
}
