% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bins.R
\name{get_gr_bin}
\alias{get_gr_bin}
\title{Create bins for a group of overlapped transcripts}
\usage{
get_gr_bin(gr, bin_size)
}
\arguments{
\item{gr}{A \code{\link[GenomicRanges]{GRanges-class}} object,
it contains transcripts from one gene or several genes
which are overlapped.}

\item{bin_size}{An integer, used to tail the gene region. Default is 50bp.}
}
\value{
A \code{\link[GenomicRanges]{GRanges-class}} object, containing
the binned region.
}
\description{
Function for generating bins for a group of overlapped transcripts.
}
