% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tq_visualization.R
\name{get_masks}
\alias{get_masks}
\title{Get target masks}
\usage{
get_masks(tq, transcripts)
}
\arguments{
\item{tq}{A \link{transcript_quantifier-class} object}

\item{transcripts}{a character vector of transcript names}
}
\value{
a \link[GenomicRanges]{GRanges-class} object
}
\description{
Retrives masks corresponding to a gene or genomic region
as a \link[GenomicRanges]{GRanges-class} object
}
