% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcript_shape_profile-class.R
\name{get_transcribed_regions}
\alias{get_transcribed_regions}
\title{Annotates transcribed regions}
\usage{
get_transcribed_regions(ranges, bw, transcribed_thresh = 1)
}
\arguments{
\item{ranges}{a \link[GenomicRanges]{GRanges-class}}

\item{bw}{the path to a bigwig file}

\item{transcribed_thresh}{the amount of signal required for a region to be
considered transcribed}
}
\description{
Returns \link[GenomicRanges]{GRanges-class} that contains all transcribed
regions within \code{ranges} with respect to \code{bw}
}
