% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tq.R
\name{get_transcripts}
\alias{get_transcripts}
\title{Get target transcripts}
\usage{
get_transcripts(
  transcript_quantifier,
  gene_name = NULL,
  chrom = NULL,
  start = NULL,
  end = NULL,
  strand = NULL
)
}
\arguments{
\item{transcript_quantifier}{A \link{transcript_quantifier-class} object}

\item{gene_name}{Name of your query gene. Only works if gene ids were given
to the \link{transcript_quantifier-class} object}

\item{chrom}{the chromosome of your query region (single value)}

\item{start}{the start position of your query region (single value)}

\item{end}{the end position of your query region (single value)}

\item{strand}{the strand specificity of your query region (default: ANY)}
}
\value{
a \link[GenomicRanges]{GRanges-class} object
}
\description{
Retrives transcripts corresponding to a gene or genomic region
as a \link[GenomicRanges]{GRanges-class} object
}
