% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcript_quantifier-class.R
\name{get_tx_id}
\alias{get_tx_id}
\title{get_tx_id}
\usage{
get_tx_id(tq)
}
\arguments{
\item{tq}{\code{\link{transcript_quantifier-class}} object}
}
\value{
a character vector of transcript ids in the same order as they are in the
\code{tq@transcripts} slot
}
\description{
Convenience function for getting transcript names from a
\code{\link{transcript_quantifier-class}} object
}
