% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ensembl_model.R
\name{get_tx_model}
\alias{get_tx_model}
\title{Get transcript model from Ensembl}
\usage{
get_tx_model(txDb_dir, dataset_name, release)
}
\arguments{
\item{txDb_dir}{A directory to save and retrieve txDb object.}

\item{dataset_name}{A dataset name used by Ensembl, like
"hsapiens_gene_ensembl".}

\item{release}{The Ensembl release version.}

\item{host_name}{The host name used to connect to Ensembl.}
}
\value{
A txDB object.
}
\description{
This function creates a txDb object if it doesn't exist locally.
}
