% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_transcript_model.R
\name{group_trancript_models}
\alias{group_trancript_models}
\title{Group transcripts}
\usage{
group_trancript_models(tx_models)
}
\arguments{
\item{tx_models}{A processed matrix without decimals in transcript model}
}
\value{
A list of grouped transcript names.
}
\description{
Group transripts if they have identical transcript model.
}
\seealso{
\code{\link{reduce_transcript_models}},
}
