% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcript_models_build_helpers.R
\name{group_transcripts}
\alias{group_transcripts}
\title{Group transcripts}
\usage{
group_transcripts(transcript_granges, distance = 0)
}
\arguments{
\item{transcript_granges}{\code{\link[GenomicRanges]{GRanges-class}}
object}

\item{distance}{the distance within which two transcripts are
considered connected}
}
\value{
A \code{\link[GenomicRanges]{GenomicRangesList-class}}
object
}
\description{
Creates groups of transcripts on the same strand based on
proximity. The groups are constructed of connected transcripts
such that a pair of transcripts are considered connected if
they are within a given distance \emph{d} of each other. The
group then consists of all transcripts that are connected to
at least one other member.
}
