% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_checkers.R
\name{is_length_two_vector}
\alias{is_length_two_vector}
\title{is_length_two_vector}
\usage{
is_length_two_vector(x)
}
\arguments{
\item{x}{a vector}
}
\value{
boolean
}
\description{
Checks that all elements in the vector is integer and the vector length is 2.
}
