% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_checkers.R
\name{is_matrix_list}
\alias{is_matrix_list}
\title{is_matrix_list}
\usage{
is_matrix_list(l)
}
\arguments{
\item{l}{a list}
}
\value{
boolean
}
\description{
Checks that all elements of list are a matrix. Will return FALSE for empty
list.
}
