% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_checkers.R
\name{is_strand_vector}
\alias{is_strand_vector}
\title{is_strand_vector}
\usage{
is_strand_vector(x)
}
\arguments{
\item{x}{a vector}
}
\value{
boolean
}
\description{
Checks that all elements are valid strand values ("+","-","*")
}
