% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tss_predictor.R
\name{label_transcripts}
\alias{label_transcripts}
\title{Generate training labels for TSS}
\usage{
label_transcripts(
  transcripts,
  cap_bigwig_plus,
  cap_bigwig_minus,
  ctrl_bigwig_plus,
  ctrl_bigwig_minus,
  ctrl_cdf_alpha = 0.01,
  radius = 150
)
}
\arguments{
\item{transcripts}{transcripts a \code{\link[GenomicRanges]{GRanges-class}}
that holds all the transcript coordinates}

\item{cap_bigwig_plus}{PRO-cap/GRO-cap data from plus strand}

\item{cap_bigwig_minus}{PRO-cap/GRO-cap data from minus strand}

\item{ctrl_bigwig_plus}{PRO-seq/GRO-seq data from plus strand}

\item{ctrl_bigwig_minus}{PRO-seq/GRO-seq data from minus strand}

\item{ctrl_cdf_alpha}{foo}

\item{radius}{The radius around the TSS in which to collect counts}
}
\value{
Return transcripts with an added active_tss column indicating if the
tss is active
}
\description{
Labels active TSS from by generating an empirical distribution
of *-cap sequencing count data across all TSS and labeling the upper tail
as active and sites with no *-cap data as inactive.
}
