% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tss_predictor.R
\name{labeled_feature_set}
\alias{labeled_feature_set}
\title{Build labeled data set}
\usage{
labeled_feature_set(
  labeled_transcripts,
  bigwig_plus,
  bigwig_minus,
  bins = 251L,
  bin_width = 21L
)
}
\arguments{
\item{labeled_transcripts}{a \code{\link[GenomicRanges]{GRanges-class}} object which
has a logical valued metadata column \code{active_tss} indicating if the TSS of the
corresponding transcript is active (TRUE -> active)}

\item{bigwig_plus}{the path to a bigwig for reads on the plus strand}

\item{bigwig_minus}{the path to a bigwig for reads on the minus strand}

\item{bins}{a single value or vector of length equal to bin_width indicating
the number of bins to generate, with the central bin centered on the
focal site, must be odd}

\item{bin_width}{a vector of one or more odd valued integers indicating the
width of the bins to be generated. If multiple value are specified, windows are
generated for all sizes specified}
}
\value{
A list with two elements containing the feature arrays for the active
and inactive TSS respectively
}
\description{
Builds and labels feature vectors for both active and inactive
TSS.
}
