% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_data.R
\name{load_txdb}
\alias{load_txdb}
\title{Load txdb from cache}
\usage{
load_txdb(file_name)
}
\arguments{
\item{file_name}{the full file name of the txdb}
}
\description{
View a list of txdbs that have been cached
}
\seealso{
\code{\link{save_txdb}},\code{\link{list_cached_txdb}}
}
