% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fit_statistics.R
\name{locus_statistics}
\alias{locus_statistics}
\title{locus_statistics}
\usage{
locus_statistics(tq, n = 1L)
}
\arguments{
\item{tq}{A \link{transcript_quantifier-class} object}

\item{n}{the number of model with non-zero abundance that can be used to
explain variance. If fewer models available it will use all available models.
Note that some models correspond to multiple transcripts due to high
similarity in co-ordinates.}
}
\value{
A table describing the variance at each locus and the variance
explained by the model(s)
}
\description{
Calculates a variety of statistics at the locus and transcript level
describing how well the model explains the variance in the data
}
