% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fit_statistics.R
\name{locus_variance_explained}
\alias{locus_variance_explained}
\title{locus_variance_explained}
\usage{
locus_variance_explained(models, abundance, data, masks)
}
\arguments{
\item{models}{the matrix of transcript models}

\item{abundance}{a matrix or vector of abundance values. When it is a
matrix each column is a set of abundance values for all models at that loci.
This is a more efficient way to calculate the variance explained at the locus
for multiple different combinations of abundance values. If it is a vector it
must have length equal to the number of models at the loci.}

\item{data}{a vector of count data}

\item{masks}{an integer vector of positions to be masked}
}
\value{
a vector of the total variance explained at the locus by all models
with one entry corresponding to each column of the aundance matrix or a
single value if abundance is a vector
}
\description{
Calculates fraction of variance in the data explained jointly by all models
at a locus
}
