% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitting.R
\name{mask_data}
\alias{mask_data}
\title{Mask data}
\usage{
mask_data(data, masks)
}
\arguments{
\item{data}{a data vector}

\item{masks}{a vector of masked indicies}
}
\description{
masks positions in data vector at indices specified in masks
vector by replacing them with zeros.
}
