% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_model_masks.R
\name{mask_transcripts}
\alias{mask_transcripts}
\title{Mask transcripts}
\usage{
mask_transcripts(transcript_models, masks)
}
\arguments{
\item{transcript_models}{A list of matricies containing transcript models}

\item{masks}{a list of}
}
\value{
A list of matrices where each row in the matrix corresponds to a
transcript and each column is a bin
}
\description{
Function for generating masks
}
