% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_checkers.R
\name{matrix_list_dim_equal}
\alias{matrix_list_dim_equal}
\title{matrix_list_dim_equal}
\usage{
matrix_list_dim_equal(l1, l2)
}
\arguments{
\item{l1}{a list}

\item{l2}{a list}
}
\value{
boolean
}
\description{
Checks that all elements of a pair of lists have the same dimensionality
}
