% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcript_shape_profile-class.R
\name{model_agreement}
\alias{model_agreement}
\title{Computes GOF statistic for groups of transcripts}
\usage{
model_agreement(tx_grps, bigwig_plus, bigwig_minus)
}
\arguments{
\item{tx_grps}{a \link[GenomicRanges]{GRangesList-class} object that contains
groups of overlapping transcripts}

\item{bigwig_plus}{the path to a bigwig for reads on the plus strand}

\item{bigwig_minus}{the path to a bigwig for reads on the minus strand}
}
\description{
Annotates transcriptes from the input \link[GenomicRanges]{GRangesList-class}
object with two additional columns: \code{percent_match} and
\code{percent_transcribed}. Within the genomic interval covered by each group
of transcripts, certain bases are predicted to be transcribed or not for each
transcript annotation. Based on the supplied bigwigs each base in the
relevant region is then annotated as containig polymerase (being transcribed)
or not. \code{percent_match} is the fraction of bases whose state, as
predicted by the transcript annotations matches the annotation from the data.
\code{percent_transcribed} is then the percent of the region predicted by the
transcript annotation to be transcribed that actually is.
}
