% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tq.R
\name{plot_model}
\alias{plot_model}
\title{Plot transcript_abundance object}
\usage{
plot_model(
  transcript_quantifier,
  gene_name = NULL,
  chrom = NULL,
  start = NULL,
  end = NULL,
  strand = NULL
)
}
\arguments{
\item{transcript_quantifier}{A \link{transcript_quantifier-class} object}

\item{gene_name}{Name of your query gene. Only works if gene ids were given
to the \link{transcript_quantifier-class} object}

\item{chrom}{the chromosome of your query region (single value)}

\item{start}{the start position of your query region (single value)}

\item{end}{the end position of your query region (single value)}

\item{strand}{the strand specificity of your query region (default: ANY)}
}
\value{
plotted tracks of transcripts and masks of your query gene
}
\description{
plot transcripts and masks
}
