% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tss_predictor.R
\name{predict_inactive_transcripts}
\alias{predict_inactive_transcripts}
\title{Return list of inactive TSS}
\usage{
predict_inactive_transcripts(tq, bigwig_plus, bigwig_minus)
}
\arguments{
\item{tq}{A \link{transcript_quantifier-class} object}

\item{bigwig_plus}{the path to a bigwig for reads on the plus strand}

\item{bigwig_minus}{the path to a bigwig for reads on the minus strand}
}
\value{
a vector of inactive transcript identifiers
}
\description{
Predicts inactive transcripts based on GRO/PRO-seq data using a p
re-trained convolutional neural net. For stability reasones, this function enforces
the use of the CPU if a TF session has not already been started. Given the small size
of the network this shouldn not impact performance significantly.
}
