% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitting.R
\name{predict_locus}
\alias{predict_locus}
\title{Predicts signal at locus level}
\usage{
predict_locus(models, abundance)
}
\arguments{
\item{models}{the matrix of transcript models}

\item{abundance}{the transcript model abundances (can be either a vector or
a matrix where each column is a vecotor of abundances)}
}
\description{
outputs prediction for read counts across entire locus
}
