% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fit_statistics.R
\name{predict_transcripts}
\alias{predict_transcripts}
\title{predict_transcripts}
\usage{
predict_transcripts(models, abundance)
}
\arguments{
\item{models}{the matrix of transcript models}

\item{abundance}{a vector of abundance values, one per transcript}
}
\value{
an matrix of predicted abundances per site (rows) and per model
(columns)
}
\description{
Predicts count data profile per transcript for a single loci
}
