% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcript_shape_profile-class.R
\name{rescale_fixed_width_position}
\alias{rescale_fixed_width_position}
\title{Tripartite linear rescaling}
\usage{
rescale_fixed_width_position(
  bin_size,
  length_out,
  linear_head_length,
  linear_tail_length,
  head_percent = 0.2,
  tail_percent = 0.2,
  short_heuristic = TRUE
)
}
\arguments{
\item{bin_size}{width of each position range that is being converted}

\item{length_out}{length of position vector to be created}

\item{linear_head_length}{total width of head region (# of bins * width)}

\item{linear_tail_length}{total width of tail region (# of bins * width)}

\item{head_percent}{fraction of [0, 1] to be occupied by head region}

\item{tail_percent}{fraction of [0, 1] to be occupied by tail region}

\item{short_heuristic}{whether sequences shorter than the sum of the head and
tail length should be accommodated by proportionally rescaling the head and
tail lengths. Otherwise an error will be thrown.}
}
\description{
Given a length, a fixed bin width, and head/tail lengths this function
generates a sequence [0, 1] that is composed of three linear sub-scales for
the head, mid, and tail region. These scales are created such that positions
<= \code{linear_head_length} are converted to a uniform sequence between
\code{[0, head_percent]} and values
\code{[bin_size * length_out - linear_tail_length, bin_size]} are converted
to a uniform sequence \code{[tail_percent, 1]} with the mid values falling
in the remaining range.
}
