% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcript_models_build_helpers.R
\name{scale_additional_masks}
\alias{scale_additional_masks}
\title{Function for scaling bins for masking additional genomic regions}
\usage{
scale_additional_masks(bins, add_mask, bin_size)
}
\arguments{
\item{bins}{A \code{\link[GenomicRanges]{GRangesList-class}} object,
each element in the list contains the bins for each corresponding group of
transcripts}

\item{add_mask}{A \code{\link[GenomicRanges]{GRanges-class}}
object records genomic regions being masked. If seqnames in add_mask are not
available in transcripts, entries associated with these seqnames will be removed.}

\item{bin_size}{width of bins}
}
\value{
A list of vectors, where each vector indicates the fractions to scale
the bins
}
\description{
Function for scaling bins for masking additional genomic regions
}
