% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bins.R
\name{seqv}
\alias{seqv}
\title{Vectorized seq.default}
\usage{
seqv(
  from = 1,
  to = 1,
  by = ((to - from)/(length.out - 1)),
  length.out = NULL,
  along.with = NULL,
  ...
)
}
\arguments{
\item{from}{the starting and (maximal) end values of the
    sequence.  Of length \code{1} unless just \code{from} is supplied as
    an unnamed argument.}

\item{to}{the starting and (maximal) end values of the
    sequence.  Of length \code{1} unless just \code{from} is supplied as
    an unnamed argument.}

\item{by}{number: increment of the sequence.}

\item{length.out}{desired length of the sequence.  A
    non-negative number, which for \code{seq} and \code{seq.int} will be
    rounded up if fractional.}

\item{along.with}{take the length from the length of this argument.}

\item{...}{arguments passed to or from methods.}
}
\description{
Version of the \code{seq} function that takes vectorized arguments
}
