% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tq.R
\name{set_datatrack_ylim}
\alias{set_datatrack_ylim}
\title{Set datatrack ylim}
\usage{
set_datatrack_ylim(data_track, ylim)
}
\arguments{
\item{data_track}{a \link[Gviz]{DataTrack-class} object}

\item{ylim}{a length two numeric vector holding the ylim boundaries.}
}
\value{
a \link[Gviz]{DataTrack-class} object
}
\description{
Set ylim of the \link[Gviz]{DataTrack-class} object
}
