% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_data.R
\name{split_path}
\alias{split_path}
\title{Split path}
\usage{
split_path(path)
}
\value{
A vector of path elements. From
\url{https://stackoverflow.com/questions/29214932/split-a-file-path-into-folder-names-vector}
}
\description{
Splits path into vector of path elements
}
