% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcript_shape_profile-class.R
\name{strip_loess}
\alias{strip_loess}
\title{Slims down loess object

Removes parts of \code{loess} object that are not required for prediction
or object printing. Edited down from code in the "strip" package.}
\usage{
strip_loess(object)
}
\arguments{
\item{object}{a loess object}
}
\description{
Slims down loess object

Removes parts of \code{loess} object that are not required for prediction
or object printing. Edited down from code in the "strip" package.
}
