% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitting.R
\name{sum_squares_grad}
\alias{sum_squares_grad}
\title{Sum of squares gradient}
\usage{
sum_squares_grad(x, models, data, transform, lambda = 0)
}
\arguments{
\item{x}{the transcript model abundances}

\item{models}{the matrix of transcript models}

\item{data}{a vector of the observed data}

\item{transform}{how to transform the data (either "identity" or "log")}

\item{lambda}{the weight of the lasso penalty (not implemented at this time)}
}
\description{
Computes gradient for sum-of-squares
}
