% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{summarize_bigwig}
\alias{summarize_bigwig}
\alias{summarize_bigwig,ANY,GRangesList-method}
\alias{summarize_bigwig,ANY,GRanges-method}
\title{Summarize bigwig}
\usage{
summarize_bigwig(
  bigwig_file,
  bins,
  summary_operation = "sum",
  autostyle_seqlevels = TRUE
)

\S4method{summarize_bigwig}{ANY,GRangesList}(
  bigwig_file,
  bins,
  summary_operation = "sum",
  autostyle_seqlevels = TRUE
)

\S4method{summarize_bigwig}{ANY,GRanges}(
  bigwig_file,
  bins,
  summary_operation = "sum",
  autostyle_seqlevels = TRUE
)
}
\arguments{
\item{bigwig_file}{a string pointing to a bigwig file}

\item{bins}{A \code{\link[GenomicRanges]{GRangesList-class}} or
\code{\link[GenomicRanges]{GRanges-class}} object where all elements in a
given GRanges (or GRangesList element) are from the same chromosome}

\item{summary_operation}{the summary opperation to apply per bin (e.g.
sum, mean, median, etc.) Defaults to "sum"}

\item{autostyle_seqlevels}{Logical. If \code{TRUE} matches the seqlevel style of the
\code{bigwig_file} amd \code{bins} object. (Default: TRUE)}
}
\value{
A list of vectors with each one corresponding to one set of bins and
each element of a vector corresponding to a bin
}
\description{
Summarizes a bigwig over a set of ranges in a
\code{\link[GenomicRanges]{GRangesList-class}}
or a \code{\link[GenomicRanges]{GRanges-class}} object
}
