% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance_tables.R
\name{transcript_abundance}
\alias{transcript_abundance}
\alias{transcript_abundance,transcript_quantifier-method}
\title{View transcript abundances}
\usage{
transcript_abundance(tq)

\S4method{transcript_abundance}{transcript_quantifier}(tq)
}
\arguments{
\item{tq}{A \link{transcript_quantifier-class} object}
}
\description{
Produces a table with estimated trancript abundances for a given
\code{\link{transcript_quantifier-class}} object
}
