% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcript_quantifier-class.R
\docType{class}
\name{transcript_quantifier-class}
\alias{transcript_quantifier-class}
\title{Class transcript_quantifier}
\description{
Class \code{transcript_quantifier} holds encodings for each transcript and
can be augmented to include data and transcript abundance estimates using the
\code{fit()} function
}
\section{Slots}{

\describe{
\item{\code{transcripts}}{a \code{\link[GenomicRanges]{GRanges-class}} that holds
all the transcript coordinates}

\item{\code{column_identifiers}}{a two element character vector that holds the column
names in \code{transcripts} for the trancript and gene identifiers
respectively. Only a column for the transcript identifier is required,
missing gene identifiers are set to NA.}

\item{\code{bins}}{a \code{\link[GenomicRanges]{GRangesList-class}} object that
records the bins for each group of transcripts}

\item{\code{bin_size}}{width of bins used to create transcript models}

\item{\code{models}}{a list of matrices with numeric values between 0 and 1
representing the fractional overlap of per transcript per bin where the rows
are the transcripts and the columns are the bins. Rownames are the
corresponding transcripts}

\item{\code{masks}}{a list of matrices with numeric values between 0 and 1
that are used to modify the models}

\item{\code{transcript_model_key}}{A three column \code{data.frame} that maps
transcripts to their group and model}

\item{\code{counts}}{a list of vectors containing the read counts per bin.
Initialized empty.}

\item{\code{upstream_polymerase_ratios}}{the log2 ratio of mean counts for each transcript
from the region immediately upstream and downstream of each TSS;
log2([500bp, 5.5kb] / [-5.5kb, -500bp])}

\item{\code{tx_gof_metrics}}{some crude per transcript goodness-of-fit metrics that are used
for heuristics in the fitting step}

\item{\code{count_metadata}}{holds information about the files the count data came from
including names and total counts}

\item{\code{model_abundance}}{A list of vectors corresponding to \code{models}
of transcript abundances. Initialized at 0.}
}}

