% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcript_quantifier-class.R
\name{transcript_quantifier_valid}
\alias{transcript_quantifier_valid}
\title{transcript_quantifier_valid}
\usage{
transcript_quantifier_valid(object)
}
\arguments{
\item{object}{a \code{transcript_quantifier} object}
}
\value{
TRUE if valid, else errors
}
\description{
Checks that a \code{transcript_quantifier} object is valid
}
