% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcript_shape_profile-class.R
\docType{class}
\name{transcript_shape_profile-class}
\alias{transcript_shape_profile-class}
\title{Class transcript_shape_profile}
\description{
Class \code{transcript_shape_profile} holds data for a set of transcripts
that are used to generate an empirical shape profile. Can be created by
calling \code{transcript_shape_profile}. A shape profile can be used with
\code{modify_model} to alter the models in a \code{transcript_quantifier}
object.
}
\section{Slots}{

\describe{
\item{\code{transcripts}}{a \link[GenomicRanges]{GRanges-class} object that contains
the ranges of transcripts used to fit the empirical shape profile.}

\item{\code{bin_size}}{bin size used to create transcript profile functions}

\item{\code{linear_head_length}}{distance into gene from TSS that is treated as the
head region for the purposes of position scaling}

\item{\code{linear_tail_length}}{distance into gene from the transcript end that is
treated as the tail region for the purposes of position scaling}

\item{\code{head_percent}}{The head region of the gene is rescaled so the position
vector that is paired with the count vector is between
\code{[0, head_percent]}}

\item{\code{tail_percent}}{The tail region of the gene is rescaled so the position
vector that is paired with the count vector is between
\code{[1 - tail_percent, 1]}}

\item{\code{shape_profile}}{a loess function that is used to predict the shape
profile}

\item{\code{shape_scale_factor}}{a scalar value to multiply the output of the loess
function by so that the profile factors predicted by the loess fit have a
median value of 1 in the central body of the gene}
}}

