% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcript_shape_profile-class.R
\name{transcript_shape_profile}
\alias{transcript_shape_profile}
\title{Creates transcript_shape_profile}
\usage{
transcript_shape_profile(
  transcripts,
  bigwig_plus = NULL,
  bigwig_minus = NULL,
  bin_size = 250,
  min_percent_transcribed = 0.8,
  min_transcript_length = 8000,
  percent_model_match = 0.8,
  linear_head_length = 10000,
  linear_tail_length = 5000,
  span = 0.4
)
}
\arguments{
\item{transcripts}{a \link[GenomicRanges]{GRanges-class} object that contains
the ranges of potential transcripts to be used to fit an empirical shape
profile. This list will be filtered using the criteria specified by the
\code{min_transcript_length}, \code{max_group_size}, and
\code{percent_model_match} parameters}

\item{bigwig_plus}{the path to a bigwig for reads on the plus strand}

\item{bigwig_minus}{the path to a bigwig for reads on the minus strand}

\item{bin_size}{width of each position range that is being converted}

\item{min_percent_transcribed}{all transcripts with pro-seq signal for less
than this percent of the transcript are removed}

\item{min_transcript_length}{all transcripts shorter than this are removed
from consideration (default: 5e3)}

\item{percent_model_match}{Based on the transcript ranges for each transcript
certain regions are expected to contain nascent RNAP signal or not. For each
transcript the percent agreement with that expectation is computed and all
transcripts below that are removed (default: 0.8).}

\item{linear_head_length}{total width of head region (# of bins * width)}

\item{linear_tail_length}{total width of tail region (# of bins * width)}

\item{span}{the parameter \eqn{\alpha} which controls the degree of
    smoothing.}
}
\description{
Creates object of \code{transcript_shape_profile}, which holds data for a set
of transcripts that are used to generate an empirical shape profile. That
shape profile can then be used with \code{modify_model} to alter the models
in a \code{transcript_quantifier} object. To get the list of transcripts to
create the shape profile a series of filters are applied and then the longest
of the remaining transcripts for each overlapping set of transcripts is
chosen. Then each transcript is binned and ...
}
