% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fit_statistics.R
\name{tx_variance_explained}
\alias{tx_variance_explained}
\title{tx_variance_explained}
\usage{
tx_variance_explained(models, abundance, data, masks, n = 1)
}
\arguments{
\item{models}{the matrix of transcript models}

\item{abundance}{a vector of abundance values, one per model}

\item{data}{a vector of count data}

\item{masks}{an integer vector of positions to be masked}

\item{n}{the number of model with non-zero abundance that can be used to
explain variance. If fewer models available it will use all available models.
Note that some models correspond to multiple transcripts due to high
similarity in co-ordinates.}
}
\value{
a vector of the variance explained per model where the names indicate
the models each entry corresponds to. When there are multiple models being
used they are seperated by a "_" delimiter
}
\description{
Calculates fraction of variance in the data explained per
transcript/combination of \code{n} transcripts at a locus.
}
