% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upstream_density_ratio.R
\name{upstream_polymerase_ratio}
\alias{upstream_polymerase_ratio}
\title{density ratio calculation}
\usage{
upstream_polymerase_ratio(
  tq,
  bigwig_plus,
  bigwig_minus,
  up_width = 5000,
  up_shift = 500,
  body_width = 5000,
  body_shift = 500
)
}
\arguments{
\item{tq}{A \link{transcript_quantifier-class} object}

\item{bigwig_plus}{the path to a bigwig for reads on the plus strand}

\item{bigwig_minus}{the path to a bigwig for reads on the minus strand}

\item{up_width}{The length of the region upstream of the TSS to calculate the
polymerase denisty in}

\item{up_shift}{how far upstream of the TSS \code{up_width} is shifted}

\item{body_width}{The length of the region downstream of the TSS to calculate the
polymerase denisty in}

\item{body_shift}{how far downstream of the TSS \code{body_width} is shifted}
}
\value{
a named vector of the polymerase ratios
}
\description{
Calculates the log2 ratio of the polymerase upstream and downstream of the TSS on the
sense strand. This ratio is used to determine if a transcript may be active. Used
during data addition to determine if a transcript specified as inactive in the
\code{fit} call should be overriden.NOTE: ratio is calculated with an additional
1 pseudocount/kb added to the numerator and denominator
}
