% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcript_shape_profile-class.R
\name{view}
\alias{view}
\alias{view,transcript_shape_profile-method}
\title{View transcript shape profile

View transcript shape model natively or scaled to a specific gene length}
\usage{
view(tsp, gene_length = NULL, ...)

\S4method{view}{transcript_shape_profile}(tsp, gene_length = NULL, ...)
}
\arguments{
\item{tsp}{A \link{transcript_shape_profile-class} object}

\item{gene_length}{A gene length, defaults to just plotting profile on}

\item{...}{for extensibility
\code{[0,1]}
axis}
}
\description{
View transcript shape profile

View transcript shape model natively or scaled to a specific gene length
}
