% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ensembl_model.R
\name{view_available_txdb}
\alias{view_available_txdb}
\title{Retrieve ensembl biomart}
\usage{
view_available_txdb(species = NULL, ensembl_version = NULL)
}
\arguments{
\item{species}{the species to download the TxDb for. Use the first letter
of the genus and the full species as one string (e.g. homo sapiens -> 
hsapiens)}

\item{ensembl_version}{The Ensembl release version.}
}
\value{
A mart object.
}
\description{
Convenience function for retrieving information about available ensembl marts
that can be turned into TxDbs.
}
