% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tss_predictor.R
\name{window_regions}
\alias{window_regions}
\title{Generate GRangesList of windows}
\usage{
window_regions(sites, bins = 21L, bin_width = 51L)
}
\arguments{
\item{sites}{a \link[GenomicRanges]{GRanges-class} containing the sites of
interest. The central site of this range will be the focal site used to
generate bins}

\item{bins}{a single value or vector of length equal to bin_width indicating
the number of bins to generate, with the central bin centered on the
focal site, must be odd}

\item{bin_width}{a vector of one or more odd valued integers indicating the
width of the bins to be generated. If multiple value are specified, windows are
generated for all sizes specified}
}
\value{
A list of vectors with each one corresponding to one set of bins and
each element of a vector corresponding to a bin
}
\description{
Generates a set of windows for each element in a
\code{\link[GenomicRanges]{GRanges-class}} object. Uses the center of each
region as the 0th co-ordinate in generating windows.
}
