% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_data.R
\name{add_data}
\alias{add_data}
\alias{add_data,transcript_quantifier-method}
\title{Add data}
\usage{
add_data(
  transcript_quantifier,
  bigwig_plus = NULL,
  bigwig_minus = NULL,
  summary_operation = "sum",
  threads = 1
)

\S4method{add_data}{transcript_quantifier}(
  transcript_quantifier,
  bigwig_plus = NULL,
  bigwig_minus = NULL,
  summary_operation = "sum",
  threads = 1
)
}
\arguments{
\item{transcript_quantifier}{A \link{transcript_quantifier-class} object}

\item{bigwig_plus}{the path to a bigwig for reads on the plus strand}

\item{bigwig_minus}{the path to a bigwig for reads on the minus strand}

\item{summary_operation}{the summary opperation to apply per bin (e.g.
sum, mean, median, etc.) Defaults to "sum"}

\item{threads}{number of threads to use}
}
\value{
A \link{transcript_quantifier-class} object with count data added.
}
\description{
adds data from bigwig to a \link{transcript_quantifier-class}
object
}
