% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_transcript_models.R
\name{create_transcript_models}
\alias{create_transcript_models}
\title{Create transcript models}
\usage{
create_transcript_models(transcripts, bins, bin_size, transcript_name_column)
}
\arguments{
\item{transcripts}{A \code{\link[GenomicRanges]{GRanges-class}}
object}

\item{bins}{A \code{\link[GenomicRanges]{GRangesList-class}} object,
each element in the list contains the bins for each corresponding group of
transcripts}

\item{bin_size}{width of bins}

\item{transcript_name_column}{A string that indicates which column in the
GRanges object contain the transcript names}
}
\value{
A list of matrices where each row in the matrix corresponds to a
bin and each column is a transcript
}
\description{
Function for generating binned transcript models given a set of bins
}
