% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitting.R
\name{fit}
\alias{fit}
\alias{fit,transcript_quantifier-method}
\title{Fit model}
\usage{
fit(transcript_quantifier, lambda = 0, threads = 1, transform = "log")

\S4method{fit}{transcript_quantifier}(transcript_quantifier, lambda = 0, threads = 1, transform = "log")
}
\arguments{
\item{transcript_quantifier}{A \link{transcript_quantifier-class} object}

\item{lambda}{the weight of the lasso penalty}

\item{threads}{number of threads to use in model fitting}

\item{transform}{how to transform the data (either "identity" or "log")}
}
\description{
Estimates trancript abundances for a given
\code{\link{transcript_quantifier-class}} object under a lasso penalty
}
