% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitting.R
\name{fit}
\alias{fit}
\alias{fit,transcript_quantifier-method}
\title{Fit model}
\usage{
fit(
  tq,
  lambda = 0,
  transform = "log",
  inactive_transcripts = NA,
  verbose = FALSE,
  heuristic_inactive_override = T
)

\S4method{fit}{transcript_quantifier}(
  tq,
  lambda = 0,
  transform = "log",
  inactive_transcripts = NULL,
  verbose = FALSE,
  heuristic_inactive_override = T
)
}
\arguments{
\item{tq}{A \link{transcript_quantifier-class} object}

\item{lambda}{the weight of the lasso penalty (not implemented at this time)}

\item{transform}{how to transform the data (either "identity" or "log")}

\item{inactive_transcripts}{a character vector listing transcripts for which abundance
values should be fixed to 0. IMPORTANT: In the case where multiple transcripts are
assigned to a single model (due to identical models at the specified bin scale) this
will be overridden if one or more transcripts assigned to the same model are active.
This will also be overidden for specific transcripts if they have 10x or more
polymerase density downstream from their TSS as they do upstream and there are no
other active transcripts in their loci.}

\item{verbose}{if TRUE shows progress bar for fitting (default: FALSE)}

\item{heuristic_inactive_override}{Boolean. If TRUE uses a series of heuristics to
re-activate some of the transcripts labeled as inactive. We reccomend you leave this
on unless your're really confident in your inactive calls. See Details for more.}
}
\description{
Estimates trancript abundances for a given
\code{\link{transcript_quantifier-class}} object under a lasso penalty
}
