% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_checkers.R
\name{is_strand_vector}
\alias{is_strand_vector}
\title{is_strand_vector}
\usage{
is_strand_vector(x, allow_star = TRUE)
}
\arguments{
\item{x}{a vector}

\item{allow_star}{is star allowed in the strand?}
}
\value{
boolean
}
\description{
Checks that all elements are valid strand values ("+","-","*"). Returns FALSE
for empty vector.
}
