% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tq_visualization.R
\name{plot_model}
\alias{plot_model}
\title{Plot transcript_abundance object}
\usage{
plot_model(
  tq,
  gene_name = NULL,
  chrom = NULL,
  start = NULL,
  end = NULL,
  strand = NULL,
  bigwig_plus = NULL,
  bigwig_minus = NULL,
  ymax_bw = NULL,
  ymax_abundance = NULL
)
}
\arguments{
\item{tq}{A \link{transcript_quantifier-class} object}

\item{gene_name}{Name of your query gene. Only works if gene ids were given
to the \link{transcript_quantifier-class} object}

\item{chrom}{the chromosome of your query region (single value)}

\item{start}{the start position of your query region (single value)}

\item{end}{the end position of your query region (single value)}

\item{strand}{the strand specificity of your query region (default: ANY)}

\item{bigwig_plus}{the path to a bigwig for reads on the plus strand}

\item{bigwig_minus}{the path to a bigwig for reads on the minus strand}

\item{ymax_bw}{maximum value of y-axis for bw files when plotting}

\item{ymax_abundance}{maximum value of y-axis for abundance when plotting}
}
\value{
plotted tracks of transcripts and masks of your query gene
}
\description{
plot transcripts and masks
}
