% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcript_models_build_helpers.R
\name{reduce_transcript_models}
\alias{reduce_transcript_models}
\title{Reduce transcript models}
\usage{
reduce_transcript_models(
  transcript_models_ls,
  transcripts,
  transcript_name_column,
  bin_operation = c("round", "floor", "ceiling")
)
}
\arguments{
\item{transcript_models_ls}{A list of matrices where each row in the matrix
corresponds to a bin and each column is a transcript.}

\item{transcripts}{a \link[GenomicRanges]{GRanges-class} object that must
contain a metadata column with a transcript id and may contain an additional
column with a gene id.}

\item{transcript_name_column}{A string that indicates which column in the
GRanges object contain the transcript names}

\item{bin_operation}{Three different modes to deal with decimals in the
transript model (due to partial overlap of the first or last exon and bins).
Either "ceiling", "floor", or "round" (default: "round").}
}
\value{
A list of matrices holding the reduced transcript models and a
dataframe holding each transcript belongs to which group and reduced model.
}
\description{
Combine identical transcript models and get non-redundent and reduced
transcript models
}
