% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_data.R
\name{split_path}
\alias{split_path}
\title{Split path}
\usage{
split_path(path)
}
\arguments{
\item{path}{file path}
}
\value{
A vector of path elements.
}
\description{
Splits path into vector of path elements
}
